chore: test hybrid with idb
